Add `init` fn. Initializes hw and returns Peripherals.
