+++ embassy-nrf-examples/src/bin/ppi.rs
@@ -19,9 +19,7 @@ use embassy_nrf::{interrupt, Peripherals};
 use gpiote::{OutputChannel, OutputChannelPolarity};
 
 #[embassy::main]
-async fn main(spawner: Spawner) {
-    let p = Peripherals::take().unwrap();
-
+async fn main(spawner: Spawner, p: Peripherals) {
     info!("Starting!");
 
     let button1 = InputChannel::new(
