+++ embassy-nrf/src/lib.rs
@@ -35,7 +35,6 @@ pub mod rtc;
 #[cfg(not(feature = "nrf52820"))]
 pub mod saadc;
 pub mod spim;
-pub mod system;
 pub mod timer;
 pub mod twim;
 pub mod uarte;
@@ -73,3 +72,88 @@ pub mod interrupt {
     pub use embassy_extras::interrupt::Priority3 as Priority;
 }
 pub use embassy_macros::interrupt;
+
+pub mod config {
+    pub enum HfclkSource {
+        Internal,
+        ExternalXtal,
+    }
+
+    pub enum LfclkSource {
+        InternalRC,
+        Synthesized,
+        ExternalXtal,
+        ExternalLowSwing,
+        ExternalFullSwing,
+    }
+
+    #[non_exhaustive]
+    pub struct Config {
+        pub hfclk_source: HfclkSource,
+        pub lfclk_source: LfclkSource,
+    }
+
+    impl Default for Config {
+        fn default() -> Self {
+            Self {
+                // There are hobby nrf52 boards out there without external XTALs...
+                // Default everything to internal so it Just Works. User can enable external
+                // xtals if they know they have them.
+                hfclk_source: HfclkSource::Internal,
+                lfclk_source: LfclkSource::InternalRC,
+            }
+        }
+    }
+}
+
+pub fn init(config: config::Config) -> Peripherals {
+    // Do this first, so that it panics if user is calling `init` a second time
+    // before doing anything important.
+    let peripherals = Peripherals::take();
+
+    let r = unsafe { &*pac::CLOCK::ptr() };
+
+    // Start HFCLK.
+    match config.hfclk_source {
+        config::HfclkSource::Internal => {}
+        config::HfclkSource::ExternalXtal => {
+            // Datasheet says this is likely to take 0.36ms
+            r.events_hfclkstarted.write(|w| unsafe { w.bits(0) });
+            r.tasks_hfclkstart.write(|w| unsafe { w.bits(1) });
+            while r.events_hfclkstarted.read().bits() == 0 {}
+        }
+    }
+
+    // Configure LFCLK.
+    match config.lfclk_source {
+        config::LfclkSource::InternalRC => r.lfclksrc.write(|w| w.src().rc()),
+        config::LfclkSource::Synthesized => r.lfclksrc.write(|w| w.src().synth()),
+
+        config::LfclkSource::ExternalXtal => r.lfclksrc.write(|w| w.src().xtal()),
+
+        config::LfclkSource::ExternalLowSwing => r.lfclksrc.write(|w| {
+            w.src().xtal();
+            w.external().enabled();
+            w.bypass().disabled();
+            w
+        }),
+        config::LfclkSource::ExternalFullSwing => r.lfclksrc.write(|w| {
+            w.src().xtal();
+            w.external().enabled();
+            w.bypass().enabled();
+            w
+        }),
+    }
+
+    // Start LFCLK.
+    // Datasheet says this could take 100us from synth source
+    // 600us from rc source, 0.25s from an external source.
+    r.events_lfclkstarted.write(|w| unsafe { w.bits(0) });
+    r.tasks_lfclkstart.write(|w| unsafe { w.bits(1) });
+    while r.events_lfclkstarted.read().bits() == 0 {}
+
+    // Init GPIOTE
+    crate::gpiote::init();
+
+    peripherals
+}
